+++ spec/jobs/reaction_cascade_job_spec.rb
@@ -1,0 +1,5 @@
+require "rails_helper"
+
+RSpec.describe ReactionCascadeJob, type: :job do
+  pending "add some examples to (or delete) #{__FILE__}"
+end
